vigil: extract `PowerManager` from `vigil.swift`
Simplify the organisation of the files by extracting the power
management logic. Restructure the API shape to be a bit more fluent in
usage.
